third_party/blink/web_tests/fullscreen/video-fixed-at-top-left.html

<!DOCTYPE html>

This test verifies videos in fullscreen mode are not affected by CSS top/left properties.
<video id="video" style="left:100px; top:100px;"></video>

<script src="full-screen-test.js"></script>
<script>
var clientRect;
var video = document.getElementById('video');

waitForEventOnceAndEnd(document, 'webkitfullscreenchange', function(){
    clientRect = video.getBoundingClientRect();
    testExpected('clientRect.left', 0)
    testExpected('clientRect.top', 0)
});

runWithKeyDown(function(){
    video.webkitRequestFullscreen()
});
</script>
